requirements.txt
acvl_utils==0.2
adan_pytorch==0.1.0
batchgenerators==0.25
dynamic_network_architectures==0.3.1
hiddenlayer==0.3
matplotlib==3.8.2
mmcv==2.0.0
mmengine==0.9.0
mmsegmentation>=2.0.0
monai==1.3.0
net==2.4
nibabel==5.2.1
nnunetv2==2.4.2
numpy==1.25.2
opencv_python_headless==4.6.0.66
pandas==1.5.2
Pillow==9.2.0
Pillow==10.3.0
pystrum==0.4
pytorch_msssim==0.2.1
Requests==2.31.0
scikit_image==0.19.3
scikit_learn==1.2.2
scipy==1.9.3
seaborn==0.13.2
SimpleITK==2.2.1
SimpleITK==2.3.1
tifffile==2022.10.10
torch==1.13.0
tqdm==4.65.2
wandb==0.16.0
torchvision==0.14.0
matplotlib==3.8.2